requirements.txt
rasa==1.10.8
transformers==2.11.0
google-api-python-client==1.8.3
spacy>=2.1,<2.2
google-cloud-storage==1.29.0
unidecode==1.1.1
torch
torchvision
google-cloud-storage
google-api-python-client
simpletransformers